backend/requirements.txt
fastapi
uvicorn
python-dotenv
groq
resend
tavily-python
httpx
pydantic
